requirements.txt
pandas
sqlalchemy
openai
python-dotenv
streamlit
